gui/requirements.txt
numpy==2.2.3
PyOpenGL==3.1.9
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
